Watcher: Fold two smoke tests into smoke-test-watcher project
In order to have less qa/ projects, this commit removes the
watcher-smoke-test-mustache and the watcher-smoke-test-painless projects
and moves the tests of both into the watcher-smoke-test projects.

This means less projects to parse when calling gradle and a shorter test
time due to being able to run everything within one elasticsearch
instance.
