--- x-pack/qa/smoke-test-watcher-with-mustache/build.gradle
@@ -1,14 +1,0 @@
-apply plugin: 'elasticsearch.standalone-rest-test'
-apply plugin: 'elasticsearch.rest-test'
-
-dependencies {
-  testCompile project(path: xpackModule('core'), configuration: 'runtime')
-  testCompile project(path: xpackModule('watcher'), configuration: 'runtime')
-  testCompile project(path: ':modules:lang-mustache', configuration: 'runtime')
-}
-
-integTestCluster {
-  setting 'xpack.security.enabled', 'false'
-  setting 'xpack.monitoring.enabled', 'false'
-  setting 'xpack.license.self_generated.type', 'trial'
-}
