+++ x-pack/qa/smoke-test-watcher/build.gradle
@@ -9,6 +9,8 @@ apply plugin: 'elasticsearch.rest-test'
 dependencies {
   testCompile project(path: xpackModule('core'), configuration: 'runtime')
   testCompile project(path: xpackModule('watcher'), configuration: 'runtime')
+  testCompile project(path: ':modules:lang-mustache', configuration: 'runtime')
+  testCompile project(path: ':modules:lang-painless', configuration: 'runtime')
 }
 
 integTestCluster {
