+++ x-pack/qa/smoke-test-watcher/src/test/java/org/elasticsearch/smoketest/WatcherTemplateIT.java
@@ -30,7 +30,7 @@ import java.util.Map;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.notNullValue;
 
-public class WatcherTemplateTests extends ESTestCase {
+public class WatcherTemplateIT extends ESTestCase {
 
     private TextTemplateEngine textTemplateEngine;
